update css and type files
