--- packages/space-opera/src/components/model_viewer_preview/model_viewer_preview_styles.css.ts
@@ -1,65 +1,0 @@
-/**
- * @license
- * Copyright 2020 Google LLC. All Rights Reserved.
- * Licensed under the Apache License, Version 2.0 (the 'License');
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an 'AS IS' BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- */
-
-
-import {css, CSSResult} from 'lit-element';
-export const styles: CSSResult = css`
-model-viewer {
-  cursor: -webkit-grab;
-  cursor: -moz-grab;
-  display: flex;
-  height: 100%;
-  width: 100%;
-  overflow: hidden;
-  position: relative;
-  user-select: none;
-  width: 100%;
-}
-
-.ScreenShotButton {
-  border-radius: 50%;
-  border: 1px #4285f4 solid;
-  bottom: 25px;
-  color: #4285f4; /* MATERIAL_COLOR_GOOGLE_BLUE_500 */;
-  position: absolute;
-  right: 25px;
-}
-
-.HelpText {
-  font-family: Roboto;
-  font-size: 20pt;
-  text-align: center;
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-}
-
-.HelpText small {
-  font-size: 12pt;
-}
-
-.ErrorText {
-  font-family: Roboto;
-  color: darkred;
-  font-size: 20pt;
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-}
-`;
